Changed behaviour of before filter
"before" now includes transactions from the stated date.
This is aligned to the balance feature (i.e. displaying balances on a certain date with all transactions causing this balance can be queried with one date).
Furthermore pagination is now more logical: The before date of one page can be the after date of the other page without missing a transaction or displaying one transaction twice.